requirements.txt
langchain==0.1.16
langchain-groq==0.1.2
python-dotenv==1.0.1
streamlit==1.33.0
requests==2.31.0
streamlit-diff-viewer